move each reducers into separate files
