+++ src/store/ducks/APP.js
@@ -1,4 +1,4 @@
-import { APP } from '../reducers/reducers';
+import { APP } from '../../store/reducers/app';
 
 export { INIT_APP } from '../constants/actionTypes';
 export { initApp } from '../actions/actions';
